Inline direct_server.py into direct.py, rename exec_backend -> direct
Merge the standalone entry point from direct_server.py into direct.py
as a main() function. Remove the now-redundant direct_server library
target and rename all exec_backend_* OCI/image targets to direct_* for
consistency with the module name.
